# backend/public/openapi.yml
openapi: 3.0.0
info:
  title: Loyalty Haven API
  description: API for managing room bookings, payments, and user authentication.
  version: 1.0.0
servers:
  - url: http://localhost:5000
    description: Local development server
security:
  - cookieAuth: []
...
paths:
  /api/auth/register:
    post:
      tags:
        - Auth
      summary: Register a new user
      description: |
        Register a new user. OTPs are sent to the provided email and phone for verification.
        A token (expires in 6 hours) is sent to the client via a cookie for authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "200":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegisterResponse"
      security: []
  /api/auth/login:
    post:
      tags:
        - Auth
      summary: Log in a user
      description: |
        Log in a user. A token (expires in 6 hours) is sent to the client via a cookie for authentication.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: User logged in successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
      security: []
...
  /api/auth/validate-credentials:
    post:
      tags:
        - Auth
      summary: Validate credentials and send OTPs
      description: |
        Send OTPs to the provided email and phone for verification during registration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                phone:
                  type: string
                email:
                  type: string
              required:
                - phone
                - email
      responses:
        "200":
          description: OTPs sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTPs sent"
      security: []
  /api/auth/send-otp:
    post:
      tags:
        - Auth
      summary: Send OTP for password reset
      description: |
        Send an OTP to the provided email for password reset.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        "200":
          description: OTP sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "OTP sent"
      security: []
  /api/auth/reset-password:
    patch:
      tags:
        - Auth
      summary: Reset user password
      description: |
        Reset the user's password using the OTP sent to their email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                oldPassword:
                  type: string
                newPassword:
                  type: string
                OTP:
                  type: string
              required:
                - email
                - oldPassword
                - newPassword
                - OTP
      responses:
        "200":
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Password reset successfully."
      security: []
...
components:
  schemas:
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        phone:
          type: string
        PhoneOTP:
          type: string
        EmailOTP:
          type: string
      required:
        - email
        - password
        - name
        - phone
        - PhoneOTP
        - EmailOTP
    RegisterResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        accountType:
          type: string
        message:
          type: string
          example: "Logged in successfully."
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        accountType:
          type: string
        message:
          type: string
          example: "Logged in successfully."
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: lh_token